+++ src/routeLoader.js
@@ -12,4 +12,7 @@ export default async function loadRoutes(app, routes) {
   const uploadRoutes = await import('./routes/upload.js');
   app.use(routes.upload, uploadRoutes.default);
 
+  const notificacionsRoutes = await import('./routes/notifications.js');
+  app.use(routes.notifications, notificacionsRoutes.default);
+
 }